Clean up warnings from static analysis
