+++ .idea/detekt.xml
@@ -6,5 +6,10 @@
         <option value="$PROJECT_DIR$/build-config/detekt.yml" />
       </set>
     </option>
+    <option name="configurationFiles">
+      <list>
+        <option value="$PROJECT_DIR$/build-config/detekt.yml" />
+      </list>
+    </option>
   </component>
 </project>
