+++ app/src/androidTest/kotlin/dev/sebastiano/bundel/FakePreferences.kt
@@ -5,7 +5,6 @@ import dev.sebastiano.bundel.preferences.schedule.TimeRangesSchedule
 import dev.sebastiano.bundel.preferences.schedule.WeekDay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.flow
 
 internal class FakePreferences : Preferences {
